requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
PyPDF2>=2.0.0
python-docx>=0.8.10
requests>=2.26.0
pydantic-settings>=2.0.3